Added state machine etc
State machine implemented in game class, I tried to implement it in the quest class but I failed, this solution will have to do. Other than that, I organized our classes into packages, and added new features to NPC class and the room class.
