+++ worldofzuul/gamefunctionality/CommandWord.java
@@ -1,10 +1,10 @@
-package worldofzuul;
+package worldofzuul.gamefunctionality;
 
 
 public enum CommandWord
 {
     GO("go"), QUIT("quit"), HELP("help"), TAKE("take"), EQUIP("equip"), CHOOSE("choose"), JOURNAL("journal"),
-    DROP("drop"), INSPECT("inspect"), APPROACH("approach"), UNKNOWN("?");
+    DROP("drop"), INSPECT("inspect"), APPROACH("approach"), INVENTORY("inventory"), UNKNOWN("?");
     
     private String commandString;
